// src/app/results/core/migrations/legacy/legacy.interfaces.ts
export namespace Legacy {
  export interface DraftResult {
    targets: DraftResultTarget[];
  }

  export interface DraftResultLine {
    amendmentReasonId?: string;
    amendmentReason?: string;
    amendmentDate?: string;
    resultLineId: string;
    originalText: string;
    resultCode: string;
    resultLevel: V1.ResultDefinitionLevel;
    parts: DraftResultLinePart[];
    choices: PromptChoice[];
    isCompleted?: boolean;
    lastSharedDate?: string;
    isEditing?: boolean;
    isDeleted: boolean;
    ignoreResult: boolean;
    noResultFound?: boolean;
    dirty?: boolean;
    ruleType?: V1.ChildResultDefinition['ruleType'];
    orderedDate: string;
    childResultDefinitions: V1.ChildResultDefinition[];
    childResultLines?: DraftResultLine[];
    atLeastOneOfChildrenNumber?: number;
    shouldHaveAtLeastOneChild?: boolean;
    conditionalMandatory: boolean;
    conditionalMandatorySelection: 'YES' | 'NO';
    delegatedPowers?: {
      userId: string;
      firstName: string;
      lastName: string;
    };
  }

  export interface DraftResultLinePart {
    value: string | PromptChoice[];
    originalText?: string;
    resultChoices?: V1.RemoteResultChoice[];
    type: V1.ResultPromptType | 'RESULT';
    state: 'UNRESOLVED' | 'RESOLVED';
    promptRef?: string;
  }

  export interface DraftResultTarget {
    applicationId?: string;
    defendantId?: string;
    draftResult: string;
    hearingId: string;
    masterDefendantId?: string;
    offenceId?: string;
    resultLines?: DraftResultLine[];
    shadowListed?: boolean;
    targetId: string;
  }

  export interface ParsedDraftResult {
    applicationId?: string;
    targetId: string;
    caseId: string;
    caseUrn?: string;
    defendantId?: string;
    masterDefendantId?: string;
    offenceId?: string;
    defendantFirstName?: string;
    defendantLastName?: string;
    shadowListed?: boolean;
    addMoreResults?: false;
    requestApproval?: false;
    results: DraftResultLine[];
    lastUpdatedAt: number;
  }

  export interface PromptChoice {
    code: string;
    label: string;
    welshLabel?: string;
    promptRef?: string;
    durationSequence?: number;
    type: string;
    hidden: boolean;
    required: boolean;
    children?: PromptChoice[];
    value?: unknown;
  }

  export interface ReusableInfo {
    reusablePrompts: V1.PromptEntry[];
    reusableResults: ResultEntry[];
  }

  export interface ResultEntry {
    shortCode: string;
    masterDefendantId: string;
    offenceId: string;
    value: string;
  }
}

// Migrated interfaces

export namespace V1 {
// ...
  export interface DraftResultRelation {
    resultLineId: string;
    ruleType: 'standalone' | 'unknown' | ChildResultDefinition['ruleType'];
    childResultLineIds: string[];
  }
// ...
  export type TargetLike<T extends object = {}> = ApplicationLike<T> | OffenceLike<T>;

  interface AbstractUnresolvedDraftResultLine {
    resultLineId: string;
    orderedDate: string;
    originalText: string;
    unresolvedParts: RemoteUnresolvedPart[];
    amendmentReason?: AmendmentReason;
    amendmentDate?: string;
  }

  export type UnresolvedDraftResultLine = TargetLike<AbstractUnresolvedDraftResultLine>;

  interface AbstractResolvedDraftResultLine
    extends Omit<AbstractUnresolvedDraftResultLine, 'unresolvedParts'> {
    deleted?: boolean;
    resultLineId: string;
    orderedDate: string;
    resultDefinitionId: string;
    label: string;
    shortCode: string;
    resultPrompts: DraftResultPrompt[];
    resultLevel: ResultDefinitionLevel;
    originalText: string;
    unresolvedParts: UnresolvedPromptPart[];
    sharedDate?: string;
    valid: boolean;
  }

  export type ResolvedDraftResultLine = TargetLike<AbstractResolvedDraftResultLine>;
// ...
  export interface PromptEntry {
    masterDefendantId?: string;
    applicationId?: string;
    cacheDataPath?: string;
    cacheable?: number;
    offenceId?: string;
    promptRef: string;
    type: ResultPromptType;
    value: unknown;
  }
// ...
  export type RemoteUnresolvedPart = RemoteUnresolvedPartForResult | RemoteUnresolvedPartForValue;
  export type RemoteResolvedPart = RemoteResolvedPartForResult | RemoteUnresolvedPartForValue;

  export interface RemoteUnresolvedPartForValue {
    type: ResultPromptType;
    state: 'UNRESOLVED';
    originalText?: string;
    value: unknown;
  }

  export interface RemoteUnresolvedPartForResult {
    state: 'UNRESOLVED';
    resultChoices: RemoteResultChoice[];
    value: string;
  }

  export interface RemoteResultChoice {
    code: string;
    label: string;
    level: ResultDefinitionLevel;
    shortCode: string;
    type: 'RESULT';
  }

  export interface RemoteResolvedPartForResult {
    type: 'RESULT';
    state: 'RESOLVED';
    value: string;
  }

  // Local part representations
  export interface UnresolvedPromptPart {
    type: ResultPromptType;
    originalText: string;
    value: unknown;
    resultPrompts: DraftResultPrompt[];
  }

  export type UnresolvedPart = RemoteUnresolvedPart | UnresolvedPromptPart;
// ...
  export type ResultPromptType = PromptChoice['type'];
// ...
  export type ResultDefinitionLevel = 'C' | 'D' | 'O';

  export interface PromptChoiceChild<P = string> {
    label: string;
    code: string;
    type: PromptChoice['type'];
    promptRef: string;
    partName: P;
    sequence: number;
    required?: boolean;
  }

  export interface ChildResultDefinition {
    code: string;
    shortCode: string;
    label: string;
    ruleType: ChildResultDefinitionRuleType;
    excludedFromResults: boolean;
    childResultCodes: string[];
  }

  export type ChildResultDefinitionRuleType = 'mandatory' | 'optional' | 'oneOf' | 'atleastOneOf';

  export interface DraftResultPrompt<V = unknown> {
    type: ResultPromptType;
    promptId: string;
    promptRef: string;
    label: string;
    welshLabel?: string;
    value: V;
    welshValue?: V;
  }
// ...
}